requirements.txt
pdfplumber==0.10.3
Pillow==10.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
